Added navigation for opening -> loading -> login.
